Move document change trigger
